+++ airbyte-integrations/connectors/source-s3/source_s3/s3file.py
@@ -33,8 +33,8 @@ from botocore import UNSIGNED
 from botocore.client import Config as ClientConfig
 from botocore.config import Config
 from botocore.exceptions import NoCredentialsError
+from source_s3.s3_utils import make_s3_client, make_s3_resource
 
-from .source_files_abstract import make_s3_client, make_s3_resource
 from .source_files_abstract.storagefile import StorageFile
